nemospeech/pyproject.toml
[project]
name = "nemospeech"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "nemo_toolkit[asr] @ git+https://github.com/NVIDIA/NeMo.git@main",
    "sounddevice",
    "soundfile",
    "requests",
    "fastapi",
    "uvicorn",
    "python-multipart",
]

[project.scripts]
nemospeech-server = "server:main"
